requirements/runtime.txt
mmcv-full>=1.2.7
numpy
prettytable
scikit-image
scipy
tensorboard
tqdm
yapf
